requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.6
djangorestframework==3.12.4
gunicorn==19.9.0
psycopg2==2.9.1
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
whitenoise==5.3.0
